+++ packages/server/src/sdk/workspace/ai/index.ts
@@ -1,6 +1,6 @@
 export * as agents from "./agents"
 export * as chatApps from "./chatApps"
+export * as configs from "./configs"
 export * as helpers from "./helpers"
 export * as rag from "./rag"
 export * as vectorDb from "./vectorDb"
-export * as models from "./models"
